Add item creation functionality
Added item creation functionality so that the app
checks to see if the db already has the item in it
or if the category does not exist, and the app
will not insert the item if that is the case.
Created queries that assist in the aforementioned
functionality. Created a basic update form page
for item updating functionality in the future.