Add delete functionality
Added delete functionality to the project so that
the user can delete a category or an item. Added
controller middleware that facilitates the
deletion of the category and its items, or a
singular item itself depending on what the user
selected. Added queries to alter the db for
deletion functionality.